Refactored Producto post method
