+++ app/Models/Producto.php
@@ -77,19 +77,119 @@ class Producto extends Model
 
         Http::withOptions([
             'verify' => false,
-            'proxy' => 'http://izdqtgr5xgbe5z:2h4gpv9haieb5u881mjjf1bio9@eu-west-static-05.quotaguard.com:9293'
+            // 'proxy' => 'http://izdqtgr5xgbe5z:2h4gpv9haieb5u881mjjf1bio9@eu-west-static-05.quotaguard.com:9293'
         ])->withHeaders([
             'apikey' => $apikey,
             'IMSURE' => 'true'
-        ])->delete(Producto::$url . '/*');
+        ])->delete(Producto::$url . '*');
 
         $response = Http::withOptions([
             'verify' => false,
-            'proxy' => 'http://izdqtgr5xgbe5z:2h4gpv9haieb5u881mjjf1bio9@eu-west-static-05.quotaguard.com:9293'
+            // 'proxy' => 'http://izdqtgr5xgbe5z:2h4gpv9haieb5u881mjjf1bio9@eu-west-static-05.quotaguard.com:9293'
         ])->withHeaders([
             'apikey' => $apikey
         ])->post(Producto::$url, $body);
 
         return $response;
+    }
+
+    public static function filter($response, $token) {
+
+        $array = array();
+
+        foreach ($response['resultado'] as $records) {
+            $product = array();
+            foreach ($records as $record) {
+                $familiaCodigo; 
+                $subfamiliaCodigo;
+                // if ($record['columna'] == 'CODART')  {
+                //     $product['referencia'] = $record['dato'];
+                // };
+                if ($record['columna'] == 'CCOART')  {
+                    $product['referencia'] = $record['dato'];
+                }
+                if ($record['columna'] == 'DESART')  {
+                    $product['nombre'] = $record['dato'];
+                }
+                if ($record['columna'] == 'FAMART')  {       
+                    $response = Http::withOptions([
+                        'verify' => false,
+                    ])->withToken($token)
+                    ->withBody(
+                        '{
+                            "ejercicio":"2022",
+                            "tabla":"F_FAM",
+                            "filtro":"CODFAM = \''.$record["dato"].'\'"
+                        }', 'application/json'
+                    )->post(FactusolProducto::$url)['resultado'];
+
+                    foreach ($response as $records) {
+                        foreach ($records as $record) {
+                            if ($record['columna'] == 'DESFAM') {
+                                $product['subfamilia'] = $record['dato'];
+                            }
+                            if ($record['columna'] == 'SECFAM') {                               
+                                $response = Http::withOptions([
+                                    'verify' => false,
+                                ])->withToken($token)
+                                ->withBody(
+                                    '{
+                                        "ejercicio":"2022",
+                                        "tabla":"F_SEC",
+                                        "filtro":"CODSEC = \''.$record["dato"].'\'"
+                                    }', 'application/json'
+                                )->post(FactusolProducto::$url)['resultado'];
+
+                                foreach ($response as $records) {
+                                    foreach ($records as $record) {
+                                        if ($record['columna'] == 'DESSEC') {
+                                            $product['familia'] = $record['dato'];
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+
+                if ($record['columna'] == 'PCOART')  {
+                    $product['precio'] = $record['dato'];
+                }
+
+                if ($record['columna'] == 'UPPART')  {
+                    $product['unidad'] = $record['dato'];
+                }
+
+                if ($record['columna'] == 'IMGART')  {
+                    $product['foto'] = $record['dato'];
+                }
+                // $product->unidadbulto = '';
+                // $product->unidadbulto2 = '';
+                // $product->unidadesxbulto = '';
+                // $product->unidadesxbulto2 = '';
+
+                //  $referencia;
+                //  $nombre;
+                //  $activo;
+                //  $subfamilia;
+                //  $familia;
+                //  $precio;
+                //  $unidad;
+                //  $unidadesxbulto;
+                //  $unidadbulto;
+                //  $unidadesxbulto2;
+                //  $unidadbulto2;
+                //  $vendounidad;
+                //  $vendobulto;
+                //  $vendobulto2;
+                //  $descripcion;
+                //  $tags;
+                //  $foto;
+                
+            }
+            // dd($product);
+            array_push($array, $product);
+        }
+        return $array;
     }
 }
